added logic to delete local repo folder

+++ app/sync.py
@@ -12,15 +12,22 @@ from db import GIT_URL, GIT_BRANCH
 def sync():
 
     #   repo = git.Repo(folder)
-    repo =  git.Repo.clone_from(GIT_URL, "./ccda_examples_repo")
-
-
     LOCAL_EXAMPLES_REPO_DIR = "./ccda_examples_repo"
 
-    shutil.rmtree(LOCAL_EXAMPLES_REPO_DIR)
+    #   delete local repo if it already exists (left over from a previous sync operation)
+    if os.path.isdir(LOCAL_EXAMPLES_REPO_DIR):
+        shutil.rmtree(LOCAL_EXAMPLES_REPO_DIR)
+
+    repo =  git.Repo.clone_from(GIT_URL, LOCAL_EXAMPLES_REPO_DIR)
+
     repo = git.Repo.clone_from(GIT_URL, LOCAL_EXAMPLES_REPO_DIR)
     repo.git.pull("origin", GIT_BRANCH)
-    parse(repo, LOCAL_EXAMPLES_REPO_DIR)
+    should_delete = parse(repo, LOCAL_EXAMPLES_REPO_DIR)
+
+    #   delete local repo afterwards
+    if should_delete:
+        shutil.rmtree(LOCAL_EXAMPLES_REPO_DIR)
+
     basedir = os.path.abspath(os.path.dirname(__file__))
 
     sections = db.sections.find().count()
